Modularize PR creation shell
